pyproject.toml
[project]
name = "thes2"
version = "0.1.0"
description = "Thesis implementation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "Django",
    "tensorflow",
    "opencv-python",
    "jdeskew",
    "Pillow"
]
